refactor : mail, product, search 코드 수정 (controller, service 역할 구분)
