--- src/main/java/com/levelUpToast/levelUpToast/mail/service/mailServiceInf/ExpireCode.java
@@ -1,7 +1,0 @@
-package com.levelUpToast.levelUpToast.mail.service.mailServiceInf;
-
-import java.util.Map;
-
-public interface ExpireCode {
-    void expireCode(String userMail, Map<String, String>codeStore);
-}
