--- src/main/java/com/levelUpToast/levelUpToast/product/service/productServiceInf/DeleteProduct.java
@@ -1,7 +1,0 @@
-package com.levelUpToast.levelUpToast.product.service.productServiceInf;
-
-import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
-
-public interface DeleteProduct {
-    void deleteProduct(Long seq) throws LevelUpToastEx;
-}
