+++ src/main/java/com/levelUpToast/levelUpToast/product/service/productServiceInf/ProductService.java
@@ -1,5 +1,6 @@
 package com.levelUpToast.levelUpToast.product.service.productServiceInf;
 
+import com.levelUpToast.levelUpToast.bodyForm.requestForm.product.ProductDeleteRequestForm;
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.bodyForm.requestForm.product.ProductListResponseForm;
 import com.levelUpToast.levelUpToast.product.domain.ResponseProductTable;
@@ -7,16 +8,17 @@ import com.levelUpToast.levelUpToast.bodyForm.requestForm.product.ProductRequest
 
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Map;
 import java.util.Optional;
 
 public interface ProductService {
-    Long registerProduct(ProductRequestForm form, Long member) throws LevelUpToastEx, SQLException;
+    Long registerProduct(ProductRequestForm form) throws LevelUpToastEx, SQLException;
 
-    void updateProduct(Optional<ResponseProductTable> original, Long seq, ProductRequestForm form) throws LevelUpToastEx, SQLException;
+    void updateProduct(Long seq, ProductRequestForm form) throws LevelUpToastEx, SQLException;
 
-    void deleteProduct(Long seq) throws LevelUpToastEx;
+    void deleteProduct(ProductDeleteRequestForm form) throws LevelUpToastEx, SQLException;
 
-    Optional<ResponseProductTable> getProduct(Long seq) throws LevelUpToastEx, SQLException;
+    Map<String, Object> getProduct(Long seq) throws LevelUpToastEx, SQLException;
 
     ArrayList<ProductListResponseForm> recommendationProduct() throws LevelUpToastEx;
