+++ src/main/java/com/levelUpToast/levelUpToast/search/service/searchServiceInf/SearchService.java
@@ -4,8 +4,9 @@ import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.bodyForm.requestForm.product.ProductListResponseForm;
 
 import java.util.List;
+import java.util.Map;
 
 public interface SearchService {
-    List<ProductListResponseForm> SearchProduct(String title, int index) throws LevelUpToastEx;
+    Map<String, Object> SearchProduct(String title, int index) throws LevelUpToastEx;
 
 }
